get and post ok

+++ livraria-backend/src/main/java/com/tqi/evolutionfullstack/livrariabackend/repository/BookOrderRepository.java
@@ -3,5 +3,5 @@ package com.tqi.evolutionfullstack.livrariabackend.repository;
 import com.tqi.evolutionfullstack.livrariabackend.model.BookOrder;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookOrderRepository extends JpaRepository<BookOrder, Integer> {
+public interface BookOrderRepository extends JpaRepository<BookOrder, Long> {
 }
