+++ livraria-backend/src/main/java/com/tqi/evolutionfullstack/livrariabackend/repository/CustomerOrderRepository.java
@@ -3,5 +3,5 @@ package com.tqi.evolutionfullstack.livrariabackend.repository;
 import com.tqi.evolutionfullstack.livrariabackend.model.CustomerOrder;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerOrderRepository extends JpaRepository<CustomerOrder, Integer> {
+public interface CustomerOrderRepository extends JpaRepository<CustomerOrder, Long> {
 }
