+++ livraria-backend/src/main/java/com/tqi/evolutionfullstack/livrariabackend/repository/CustomerRepository.java
@@ -1,7 +1,9 @@
 package com.tqi.evolutionfullstack.livrariabackend.repository;
 
+import com.tqi.evolutionfullstack.livrariabackend.dto.NewCustomer;
 import com.tqi.evolutionfullstack.livrariabackend.model.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepository extends JpaRepository<Customer, Integer> {
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
+
 }
